+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - use convenience function to obtain organizational unit merged ids
 - lock file maintenance (2026-01-12)
 - remove mailpit path for CI testing
+- datenkompass: publish activities by unit, filter and mapping update
 
 ### Deprecated